Use explicit widths when defining protocol fields in messages
Best practices for network programming is to use types from <stdint.h>
which have explicit numbers of bits.

Note: this change breaks the API in that the session_id and datalength
fields of the Tacacs+ header are now signed.  (The addition of the
arg_len[0] members to acct_reply, author, and author_reply, or of msg[0]
to authen_cont and authen_reply, should not affect the API or ABI.)