Don't compute pad all at once when we can compute it incrementally
This saves us having to marshall data and allocate a buffer for
the entire pad (bitstream cipher).  We only need it in blocks of
16 bytes (the size of the MD5 digest), so let's compute it piecemeal
as we need it.

This has the added benefit of avoiding any calls to malloc() which
might be result in system calls (i.e. sbrk() to extend the heap).