Add option to specify modem port, etc.
Extracting the tty name or port name from the controlling terminal
of a Unix process is a questionable assumption.  Further, for
automated testing, you might want the reproducibility of being
able to explicitly set the terminal name.  To get the same
functionality as before, one can pass in "tacc ... -y `tty`" as
an extra argument.