Add the notion of 'idling' a session when we get a response
A session is 'busy' when a request has been sent, and the
response not yet received.  As soon as the response is received,
the session re-enters the 'idle' state (and any associated read
timeouts can be cleared).

We also clear timeouts when a session enters connected state,
but only if we hadn't pre-queued the first request.