attrib: Fix overflow issue in attr length check
In tac_add_attrib_pair() the lengths of the name and value were
stored in unsigned char variables, which can represent a maximum
value of 255.

In practice this caused the total_len > 255 conditional to be false
in the majority of cases, since the length overflowed and wrapped to
a value < 255.