attrib: tac_add_attrib funcs return int
tac_add_attrib() and tac_add_attrib_pair() now return a value < 0 if
the attribute could not be added because it was greater than 255
characters in length.

This allows the caller to detect when such a failure occurs.

Bump the library version since this is an API change, and accordingly
replace the libtac2 Debian package with libtac3.