attrib: Add funcs which truncate to max length
Add two new functions:
- tac_add_attrib_truncate()
- tac_add_attrib_pair_truncate()

These functions behave identically to the existing add_attrib
functions except they will truncate value as required in order to
keep the total length of the attribute equal to or less than 255
(the max attribute length supported by TACACS+).

This is useful for eg. a command accounting record which can be
performed on a best effort basis, if the attribute length or count
is exceeded.