Fix RNG when startup occurred recently
The random number generator function (magic()) uses the 'getrandom()'
function. This function relies on the /dev/urandom pool, which may not
be initialized when the system has only recently started up. If this
function is called before the pool is initialized, then the program will
abort.

Added a check for if this case occurs, and in that case the OpenSSL
functions are used instead if they are present, otherwise the random()
function is used.