Fix double free issue in authen_r.c
As part of tac_authen_read_timeout(), in line 141, the msg variable pointer is saved in re->msg.
The immediate next line frees this msg pointer which causes re->msg to become a dangling pointer.
In pam_sm_authenticate(), which is the caller for the above mentioned function, this pointer is stored as part of arep.msg and is again freed in line 552.
This causes the below message to be seen:
free(): double free detected in tcache 2
Aborted (core dumped)

So, fix is to not free the msg pointer in tac_authen_read_timeout().

Signed-off-by: Rahul Sreeram <rahul.sreeram@gmail.com>